Fix warnings

+++ Ukemochi-Engine/src/Ukemochi-Engine/Math/Transformation.cpp
@@ -43,14 +43,14 @@ namespace Ukemochi
 
 	void Transformation::IncreaseScale(Transform& trans)
 	{
-		trans.scale += Vec2{ SCALE_FACTOR, SCALE_FACTOR } * g_FrameRateController.GetDeltaTime();
+		trans.scale += Vec2{ SCALE_FACTOR, SCALE_FACTOR } * static_cast<float>(g_FrameRateController.GetDeltaTime());
 		trans.scale.x = clamp(trans.scale.x, MIN_SCALE, MAX_SCALE);
 		trans.scale.y = clamp(trans.scale.y, MIN_SCALE, MAX_SCALE);
 	}
 
 	void Transformation::DecreaseScale(Transform& trans)
 	{
-		trans.scale -= Vec2{ SCALE_FACTOR, SCALE_FACTOR } * g_FrameRateController.GetDeltaTime();
+		trans.scale -= Vec2{ SCALE_FACTOR, SCALE_FACTOR } *static_cast<float>(g_FrameRateController.GetDeltaTime());
 		trans.scale.x = clamp(trans.scale.x, MIN_SCALE, MAX_SCALE);
 		trans.scale.y = clamp(trans.scale.y, MIN_SCALE, MAX_SCALE);
 	}
